Remove AssetViewModel dependency from HomeViewModel
