+++ neo-gui/UI/Home/AssetsView.xaml
@@ -1,9 +1,15 @@
-﻿<ListView x:Class="Neo.UI.Home.AssetsView"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:properties="clr-namespace:Neo.Properties"
-        ItemsSource="{Binding Assets}" SelectedItem="{Binding SelectedAsset}"
-        Loaded="AssetsView_Loaded" MouseDoubleClick="AssetList_DoubleClick">
+﻿<ListView 
+    x:Class="Neo.UI.Home.AssetsView"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="clr-namespace:Neo.UI.Home"
+    xmlns:markupExtensions="clr-namespace:Neo.UI.MarkupExtensions"
+    xmlns:properties="clr-namespace:Neo.Properties"
+    ItemsSource="{Binding Assets}" 
+    SelectedItem="{Binding SelectedAsset}"
+    DataContext="{markupExtensions:DataContextBinding ViewModel=local:AssetsViewModel}"
+    MouseDoubleClick="AssetList_DoubleClick">
+    
     <ListView.View>
         <GridView>
             <GridViewColumn Header="{x:Static properties:Strings.Asset}" Width="120" DisplayMemberBinding="{Binding Name}" />
