+++ neo-gui/UI/Home/AssetsView.xaml.cs
@@ -1,5 +1,4 @@
 ﻿using System.Diagnostics;
-using System.Windows;
 using System.Windows.Input;
 using Neo.Properties;
 
@@ -15,19 +14,6 @@ namespace Neo.UI.Home
         public AssetsView()
         {
             InitializeComponent();
-        }
-
-        private void AssetsView_Loaded(object sender, RoutedEventArgs e)
-        {
-            this.AttachViewModel();
-        }
-
-        private void AttachViewModel()
-        {
-            // Check if view model has already been attached
-            if (this.viewModel != null) return;
-
-            this.viewModel = this.DataContext as AssetsViewModel;
         }
 
         private void AssetList_DoubleClick(object sender, MouseButtonEventArgs e)
