+++ neo-gui/UI/Messages/AddAssetMessage.cs
@@ -1,0 +1,12 @@
+﻿namespace Neo.UI.Messages
+{
+    public class AddAssetMessage
+    {
+        public AssetItem AssetItem { get; private set; }
+
+        public AddAssetMessage(AssetItem assetItem)
+        {
+            this.AssetItem = assetItem;
+        }
+    }
+}
