+++ neo-gui/UI/Messages/ClearAssetsMessage.cs
@@ -1,0 +1,6 @@
+﻿namespace Neo.UI.Messages
+{
+    public class ClearAssetsMessage
+    {
+    }
+}
