+++ neo-gui/UI/Messages/UpdateAssetsBalanceMessage.cs
@@ -1,10 +1,10 @@
 ﻿namespace Neo.UI.Messages
 {
-    public class WalletBalanceChangedMessage
+    public class UpdateAssetsBalanceMessage
     {
         public bool BalanceChanged { get; private set; }
 
-        public WalletBalanceChangedMessage(bool balanceChanged)
+        public UpdateAssetsBalanceMessage(bool balanceChanged)
         {
             this.BalanceChanged = balanceChanged;
         }
