+++ neo-gui/neo-gui.csproj
@@ -385,11 +385,14 @@
     <Compile Include="UI\AccountType.cs" />
     <Compile Include="UI\MarkupExtensions\DataContextBindingExtension.cs" />
     <Compile Include="UI\Messages\AccountBalanceChangedMessage.cs" />
+    <Compile Include="UI\Messages\AddAssetMessage.cs" />
     <Compile Include="UI\Messages\ClearAccountsMessage.cs" />
+    <Compile Include="UI\Messages\ClearAssetsMessage.cs" />
     <Compile Include="UI\Messages\EnableMenuItemsMessage.cs" />
     <Compile Include="UI\Messages\LoadWalletAddressesMessage.cs" />
     <Compile Include="UI\Messages\RestoreContractsMessage.cs" />
     <Compile Include="UI\Messages\UpdateAcountListMessage.cs" />
+    <Compile Include="UI\Messages\UpdateAssetsBalanceMessage.cs" />
     <Compile Include="UI\Messages\WalletBalanceChangedMessage.cs" />
     <Compile Include="UI\TxOutListBox.xaml.cs">
       <DependentUpon>TxOutListBox.xaml</DependentUpon>
